Split-out conversion and helper functions
